# src/swagger/swagger-config/swagger.yml
openapi: 3.0.0
paths:
...
  /order:
    post:
      operationId: OrderController_createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createOrderDTo'
      responses:
        '200':
          description: Creats the new Order
          content:
            application/json:
              examples:
                succes: &ref_0
                  description: Gives the order details
                  value:
                    succes: true
                    data:
                      customerName: premchandu
                      tableNumber: 20
                      order_id: 14
                      date:
                        createdDate: '2023-08-24T08:24:49.956Z'
                        updatedDate: '2023-08-24T08:24:49.956Z'
        '400':
          description: Invalid Requests
          content:
            application/json:
              examples:
                succes: *ref_0
                TableNumberNull:
                  description: Given Table is not entered
                  value:
                    message: Table number should be number
                QuantityNull:
                  description: This error occurs when quanitty is null
                  value:
                    message:
                      - quantity should not be empty
                    error: Bad Request
                MenuItemNull:
                  description: This error occurs when quanitty is null
                  value:
                    message:
                      - menuItem should not be empty
                    error: Bad Request
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
      tags: &ref_1
        - Orders
      security: &ref_2
        - bearer: []
  /order/{id}:
    get:
      operationId: OrderController_getOrderDetailsById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Gives the orders details by given id
          content:
            application/json:
              examples:
                Success:
                  description: Gives the order details
                  value:
                    succes: true
                    data:
                      order_id: 11
                      customerName: ramanjenulu
                      tableNumber: 0
                      orderStatus: completed
                      paymentStatus: approved
                      orderItems:
                        - orderItem_id: 17
                          quantity: 3000
                          menuitems:
                            menuitem_id: 4
                            menu_itemname: Pancakes
                            price: 10
                            tax: 5
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
        '404':
          description: Order with id not found
          content:
            application/json:
              examples:
                NotFound:
                  description: Gives the order details
                  value:
                    message: Invalid_Id_Order_with_given_id_is_not_available
      tags: *ref_1
      security: *ref_2
  /order/itemquantity{id}:
    put:
      operationId: OrderController_updateOrderQuantity
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateOrderDto'
      responses:
        '200':
          description: >-
            Updates the quantity with given menuitem name and quantity to be
            placed
          content:
            application/json:
              examples:
                success:
                  description: updates the menuItem and returns menuitem
                  value:
                    success: true
                    data:
                      succes: true
                      data:
                        orderItem_id: 17
                        quantity: 31
                        menuitems:
                          menuitem_id: 4
                          menu_itemname: Pancakes
                          price: 10
                          tax: 5
        '400':
          description: Id is not found with given OrderItemId
          content:
            application/json:
              examples:
                InvalidId:
                  description: >-
                    This error occurs when name with given Orderitemid is not
                    found
                  value:
                    message: Id_with_given_OrderItemId_is_not_avalaible
                    error: Bad Request
                InvalidMenuItem:
                  description: This error occurs when menuItemName  is not found
                  value:
                    message: Invalid_Menuitem_name
                    error: Bad Request
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
      tags: *ref_1
      security: *ref_2
  /order/{orderItemId}:
    delete:
      operationId: OrderController_deleteMenuItem
      parameters:
        - name: menuItemid
          required: true
          in: path
          schema:
            type: number
      responses:
        '400':
          description: Occurs when order is deleted
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Gives the order details
                  value:
                    message: Invalid_Id_Order_with_given_id_is_not_available
      tags: *ref_1
      security: *ref_2
  /order/{Orderid}:
    delete:
      operationId: OrderController_deleteOrder
      summary: ''
      description: Deletes the odrer with given orderid
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Delets the order with given OrderId
          content:
            application/json:
              examples:
                success:
                  description: Deletes the Order
                  value:
                    message: Order_with_Id__has_been_deleted_successfuly
        '400':
          description: Occurs when order is deleted
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Gives the order details
                  value:
                    message: Invalid_Id_Order_with_given_id_is_not_available
      tags: *ref_1
      security: *ref_2
  /order/{id}/addItem:
    put:
      operationId: OrderController_addMenuItem
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateOrderDto'
      responses:
        '200':
          description: >-
            Updates the quantity with given menuitem name and quantity to be
            placed
          content:
            application/json:
              examples:
                success:
                  description: updates the menuItem and returns menuitem
                  value:
                    success: true
                    data:
                      succes: true
                      data:
                        orderItem_id: 17
                        quantity: 31
                        menuitems:
                          menuitem_id: 4
                          menu_itemname: Pancakes
                          price: 10
                          tax: 5
        '401':
          description: Unauthorized
          content:
            application/json:
              examples:
                AcceswithoutAuthorization:
                  description: ' If the client attempts to access certain book-related endpoints without providing any authentication credentials'
                  value:
                    message: Authentication_is_required_to_access_the_resource.
                ExpiredAuthenticationToken:
                  description: >-
                    When the client sends an authentication token that is
                    invalid, expired, or not recognized by the server
                  value:
                    message: client's_token_is_not_valid_for_the_requested_operation
        '403':
          description: Forbidden
          content:
            application/json:
              examples:
                InsufficientPrivileges:
                  description: ' If the client provides authentication credentials, but the credentials do not grant sufficient privileges to access a specific book-related operation'
                  value:
                    message: >-
                      client_does_not_have_the_required_authorization_to_perform_the_requested_action.
                RoleBasedAccessControl:
                  description: >-
                    If the API uses role-based access control, the server may
                    respond with a 403 status code when a user with a certain
                    role tries to access an operation that requires a higher
                    role or specific permissions.
                  value:
                    message: >-
                      User_with_this_role_is_not_permitted_to_access_the_resource
      tags: *ref_1
      security: *ref_2
  /order/approved/{id}:
    patch:
      operationId: OrderController_updatePaymentStatus
      summary: ''
      description: Update the payment status as approved using the given OrderId
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Updates the order and payment status
          content:
            application/json:
              examples:
                success:
                  description: Updates the payment status
                  value:
                    message: Order_with_Id_payment_has_been_updated_successfuly
        '400':
          description: Occurs when order is deleted
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Gives the order details
                  value:
                    message: Invalid_Id_Order_with_given_id_is_not_available
      tags: *ref_1
      security: *ref_2
  /order/declined/{id}:
    patch:
      operationId: OrderController_updatePaymentdeStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '400':
          description: Occurs when order is deleted
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Gives the order details
                  value:
                    message: Invalid_Id_Order_with_given_id_is_not_available
      tags: *ref_1
      security: *ref_2
...
info:
  title: chandu
  description: Your API Description
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    OrderItemDTo:
      type: object
      properties:
        menuItemId:
          type: number
          example: 1
        quantity:
          type: number
          example: 2
      required:
        - menuItemId
        - quantity
    createOrderDTo:
      type: object
      properties:
        customerName:
          type: string
          example: Sukesh
        items:
          example:
            - menuItemId: 3
              quantity: 1
            - menuItemId: 4
              quantity: 1
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTo'
      required:
        - customerName
        - items
    updateOrderDto:
      type: object
      properties:
        menuItem:
          type: string
          example: Pancakes
        quantity:
          type: number
          example: 10
      required:
        - menuItem
        - quantity
...